[spreader][teams][test] Refactoring. Extracted abstract part of test. The first step on the way approvals test for teams.
