+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/web/rest/AbstractTeamControllerTest.java
@@ -1,0 +1,147 @@
+package com.codenjoy.dojo.web.rest;
+
+import com.codenjoy.dojo.services.Deal;
+import com.codenjoy.dojo.services.GameServiceImpl;
+import com.codenjoy.dojo.services.GameType;
+import com.codenjoy.dojo.services.mocks.FirstGameType;
+import com.codenjoy.dojo.services.mocks.SecondGameType;
+import com.codenjoy.dojo.services.mocks.ThirdGameSettings;
+import com.codenjoy.dojo.services.mocks.ThirdGameType;
+import com.codenjoy.dojo.services.multiplayer.GameRoom;
+import com.codenjoy.dojo.web.rest.pojo.PTeam;
+import org.json.JSONArray;
+import org.junit.Before;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+
+import java.util.*;
+
+import static com.codenjoy.dojo.stuff.SmartAssert.assertEquals;
+import static com.codenjoy.dojo.utils.JsonUtils.toStringSorted;
+import static java.util.stream.Collectors.*;
+import static java.util.stream.Collectors.joining;
+
+public class AbstractTeamControllerTest extends AbstractRestControllerTest  {
+
+    public static final String game = "third";
+    public static final String ip = "ip";
+    public static final String room = "test";
+
+    protected ThirdGameSettings settings;
+    protected ThirdGameType type;
+
+    @TestConfiguration
+    public static class ContextConfiguration {
+        @Bean("gameService")
+        public GameServiceImpl gameService() {
+            return new GameServiceImpl(){
+                @Override
+                public Collection<? extends Class<? extends GameType>> findInPackage(String packageName) {
+                    return Arrays.asList(
+                            FirstGameType.class,
+                            SecondGameType.class,
+                            ThirdGameType.class);
+                }
+            };
+        }
+    }
+
+    @Before
+    public void setUp() {
+        super.setUp();
+
+        players.removeAll();
+        rooms.removeAll();
+        registration.removeAll();
+
+        // when changing teams, the current state is
+        // saved to the database (there is a TeamId).
+        // It should also be deleted between test runs.
+        saves.removeAllSaves();
+
+        // this is how we get the room settings
+        type = (ThirdGameType) games.getGameType(game);
+        settings = (ThirdGameSettings) rooms.create(room, type).getSettings();
+
+        // We want to remember every field ever created,
+        // in order to count the indices. So there we want to delete it.
+        type.clear();
+
+        asAdmin();
+
+        // for all tests
+        settings.playersAndTeamsPerRoom(4, 2);
+    }
+
+    public void givenPl(PTeam... teams) {
+        for (PTeam team : teams) {
+            for (String playerId : team.getPlayers()) {
+                register(playerId, ip, room, game);
+            }
+        }
+        teamService.distributePlayersByTeam(room, Arrays.asList(teams));
+        removeDeadFields();
+    }
+
+    /**
+     * Reset uninformative indexes that
+     * could have been created during the
+     * registration process, changing commands.
+     */
+    private void removeDeadFields() {
+        type.fields().clear();
+        type.fields().addAll((List)deals.rooms().values().stream()
+                .map(room -> room.field())
+                .distinct()
+                .collect(toList()));
+    }
+
+    public void asrtTms(String expected) {
+        String actual = deals.all().stream()
+                .collect(groupingBy(Deal::getTeamId, TreeMap::new, toSet()))
+                .entrySet().stream()
+                .map(entry -> {
+                    Integer teamId = entry.getKey();
+                    String players = entry.getValue().stream()
+                            .map(Deal::getPlayerId)
+                            .sorted()
+                            .collect(joining(","));
+                    return String.format("[%d: %s]\n", teamId, players);
+                })
+                .collect(joining());
+        assertEquals(expected, actual);
+    }
+
+    public void asrtFld(String expected) {
+        Collection<GameRoom> rooms = deals.rooms().get(room);
+
+        Map<String, Integer> playersTeams = deals.all().stream().
+                collect(toMap(deal -> deal.getPlayerId(),
+                        deal -> deal.getTeamId()));
+
+        String actual = type.fields().stream()
+                .map(field -> {
+                    String players = rooms.stream()
+                            .filter(room -> room.field() == field)
+                            .flatMap(room -> room.players().stream())
+                            .map(gamePlayer -> deals.get(gamePlayer).get().getPlayerId())
+                            .map(id -> String.format("%s(t%s)", id, playersTeams.get(id)))
+                            .collect(joining(", "));
+                    return String.format("[f%s: %s]", type.fields().indexOf(field), players);
+                })
+                .collect(joining("\n"));
+
+        assertEquals(expected, actual);
+    }
+
+    public void callGet(PTeam... teams) {
+        String expected = new JSONArray(Arrays.asList(teams)).toString();
+        String actual = new JSONArray(get("/rest/team/room/" + room)).toString();
+        assertEquals(expected, actual);
+    }
+
+    public void callPost(PTeam... teams) {
+        post(202, "/rest/team/room/" + room,
+                toStringSorted(Arrays.asList(teams)));
+    }
+}
